방 나가기 및 엔티티 수정

See merge request s09-final/S09P31A701!18

+++ honey-back/ggulswriting/src/main/java/com/thirty/ggulswriting/room/dto/request/RoomParticipateReqDto.java
@@ -17,6 +17,6 @@ import lombok.NoArgsConstructor;
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class RoomParticipateReqDto {
 	@NotNull
-	private Long roomId;
+	private int roomId;
 
 }
